feat(app): fail api button
- Add `Fail API` button to the page which will return a error response
